This commit "closes the loop" on the revamped WebCL implementation. Firstly, it contained the command-buffer implementation of OpenCL, as contained in the ocl_* file. These files were generated by a code generator that takes in cl.h and outputs all command-buffer facilities, include the client functions, the messages, the message map from messages to service functions, and the service functions. Secondly, a few issues on the V8 side (in third_party/WebKit/Source/modules/webcl) were fixed in the process of hooking up the command-buffer client functions.
With this commit, all necessary components of WebCL in Chromium are in place.
webcl.getPlatforms() returns correct results (client/service functions for
clGetPlatformIDs(), which is used by this call, have been traced through
to make sure that the command buffer facilities are good), but testing has
not gone beyond this single function. The next step is more comprehensive
testing and debugging with real WebCL JS examples.