add extension support KHR_fp64.
It appears that AMD driver reports supported extensions on a platform as the INTERSECTION of extension supported by devices. Therefore very few extensions are reported as available on the platform. Even CL-GL interop is not reported as available. The platform extensions should be the UNION of extensions supported by different devices.